pyproject.toml
[tool.poetry]
name = "kdrive-download-category"
version = "0.1.0"
description = ""
authors = ["Denis Metrailler"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
requests = "^2.32.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
